Specific routers have been created for users and threads in the API. A specific connector has also been created for the database.
